requirements.txt
gspread
oauth2client
tweepy
requests
python-dotenv
schedule
